refactor: clean up `Graph` component
